Theme Options: Don't hide 'Try & Customize' if user can't 'edit_theme_options'
Since the `sites` subtree isn't necessarily populated with data for all of a user's sites at the time this selector is evaluated, I'm removing it.
From a UX POV, it's better to have a correct sites list available in the dropdown, even when the customizer might not actually work for some of those sites because the user lacks privileges.
This should be revisited once the site selector can reliably query required data for the selectors we're using. (Right now, we're advising agains using e.g. `<QuerySites allSites />`, as it duplicates data found in `sites-list`, and would persist that data, which can be quite a lot -- and our persisted state tree is a bit bloated already.)

This selector is different than others (and more problematic) since its default value will cause the option to be hidden instead of shown.